[firewall] support Netfilter FORWARD hook
* [firewall] support Netfilter FORWARD hook filtering
* [nftables] drop input rules moved under exported method DropIPv4Input
* [firewall] drop input chain is now optional. Can be enabled via flags
* [firewall] move errors under firewall/errors.go
* [nftables] remove logger from the fields of the struct

+++ cmd/nettrust.go
@@ -72,7 +72,13 @@ func main() {
 
 	// Firewall
 	fw, err := firewall.NewFirewall(
-		config.FirewallType, tableNameOutput, chainNameOutput, logger)
+		config.FirewallBackend,
+		config.FirewallType,
+		tableNameOutput,
+		chainNameOutput,
+		config.FirewallDropInput,
+		logger,
+	)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -149,9 +155,11 @@ func main() {
 		if err != nil {
 			log.Fatal(err)
 		}
-		err = fw.DeleteChain(chainNameInput)
-		if err != nil {
-			log.Fatal(err)
+		if config.FirewallDropInput {
+			err = fw.DeleteChain(chainNameInput)
+			if err != nil {
+				log.Fatal(err)
+			}
 		}
 		err = fw.DeleteTable(tableNameOutput)
 		if err != nil {
